requirements.txt
python-dotenv
cloudinary
tqdm
requests
tweepy